requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.126.0
Flask==3.1.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psutil==7.1.3
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.39.0
Werkzeug==3.1.4
Pillow>=10.0.0
psycopg2-binary>=2.9.9
